refactored search controller and changed a thumbnail name because it was not showing
