+++ config/routes.rb
@@ -18,5 +18,5 @@ Rails.application.routes.draw do
   resources :champions, only: [:show]
   resources :account_activations, only: [:edit]
   resources :password_resets,     only: [:new, :create, :edit, :update]
-  get 'search_friends', to: 'profiles#search'
+  get 'search_friends', to: 'search#search'
 end
